Separate collapse card component to file
